bashrc: add an alias to remove neovim plugins thingies
Sometimes the plugins act up so I need to remove them altogether and
re-install them. This alias helps with that. It rarely needs to be done
but is now there when I need, instead of running `find <> | xargs rm`
which might pose a real danger to my $HOME if I am not careful and I am
rarely careful.