bashrc: rename the `sudo` alias to `envsudo`
The `sudo` alias on non-NixOS machines preserves the PATH environment
variable so that Nix binaries made available by home-manager in an
updated PATH can be used with a `sudo` prefix for quick escalation of
privilege. This would cause `sudo -i` to fail on said systems and as a
result, instead of performing shenanigans with a function call that
evaluates the first argument passed to `sudo`, rename the alias itself
to something else, in this case `envsudo`.